+++ src/main/java/org/crs/se2035jv_anhndhe200028_carrentingsystem/repository/CarProducerRepository.java
@@ -3,6 +3,6 @@ package org.crs.se2035jv_anhndhe200028_carrentingsystem.repository;
 import org.crs.se2035jv_anhndhe200028_carrentingsystem.entity.CarProducer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarProducerRepository extends JpaRepository<CarProducer, Long> {
+public interface CarProducerRepository extends JpaRepository<CarProducer, Integer> {
     CarProducer findByProducerName(String producerName);
 }
