+++ src/main/java/org/crs/se2035jv_anhndhe200028_carrentingsystem/repository/CarRentalRepository.java
@@ -3,5 +3,5 @@ package org.crs.se2035jv_anhndhe200028_carrentingsystem.repository;
 import org.crs.se2035jv_anhndhe200028_carrentingsystem.entity.CarRental;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarRentalRepository extends JpaRepository<CarRental, Long> {
+public interface CarRentalRepository extends JpaRepository<CarRental, Integer> {
 }
